[TM-2628] Remove the JobsSlice.
It was kind of broken - it would really only display the status information for the most recent pending job, and it was being relied on in a weird way for job polling. Instead, simply rely on the cached delayed jobs and make sure to fetch once when we load the site with a logged in user.